Assignment_3/pyproject.toml
[project]
name = "ass3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "pillow>=12.2.0",
    "pyyaml>=6.0.3",
    "roboflow>=1.3.1",
    "ultralytics>=8.4.36",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchvision = [{ index = "pytorch-cu124" }]

[tool.uv]
environments = ["sys_platform == 'win32'"]
